requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Pydantic
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Document parsing
PyMuPDF>=1.23.0
python-docx>=1.1.0

# Tokenization for chunking
tiktoken>=0.5.0

# Voyage AI for embeddings and reranking
voyageai>=0.2.0

# Qdrant vector database
qdrant-client>=1.7.0

# Google Gemini
google-generativeai>=0.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
